Fix wrong return type for Value property in IConstIterator<T, U> interface

+++ Source/Tests/RangeIt.Tests/Iterators/Interfaces/IConstIterator'2_Tests.cs
@@ -67,7 +67,7 @@
 
             uriTemplatePropertyInfo.CanRead.Should().BeTrue();
             uriTemplatePropertyInfo.CanWrite.Should().BeFalse();
-            uriTemplatePropertyInfo.PropertyType.Should().Be(typeof(int));
+            uriTemplatePropertyInfo.PropertyType.Should().Be(typeof(float));
         }
     }
 }
